Fix crashes in BLE::OnBleEndPointConnectionClosed
In BLE::OnBleEndPointConnectionClosed, the endpoint is already closed.
We may close it again in the Transport::BLE destructor, which can crash.
Clear the endpoint in OnBleEndPointConnectionClosed to avoid this case.

In addition, the OnRendezvousError callback in the typical case stops
the rendezvous process which immediately destroys up the transport. We
therefore can't call OnRendezvousConnectionClosed afterwards; |this| is
already freed and we just need to return.